Refactor: rename DashboardCardsTracker to CardsTracker
